✨ Account deletion impl.

+++ src/Application/CQRS/User/DeleteUser/DeleteUserCommandHandler.cs
@@ -1,3 +1,5 @@
+using SharedKernel.Enums;
+
 namespace Application.CQRS.User.DeleteUser;
 
 public sealed class DeleteUserCommandHandler(IUnitOfWork unitOfWork) : ICommandHandler<DeleteUserCommand, Result>
@@ -11,9 +13,17 @@ public sealed class DeleteUserCommandHandler(IUnitOfWork unitOfWork) : ICommandH
             return Result.Failure(
                 ErrorFactory.NotFound($"User with ID {command.UserId} not found.")
             );
-
+        
+        var orders = await _unitOfWork.OrderRepository.OrdersByUserAsync(command.UserId);
+        if (orders.Count > 0 && orders.Any(o => o.Status == OrderStatus.Pending))
+        {
+            return Result.Failure(
+                ErrorFactory.Conflict("User cannot be deleted while having pending orders.")
+            );
+        }
+        
+        _unitOfWork.OrderRepository.DeleteRange(orders);
         _unitOfWork.CartRepository.Delete(await _unitOfWork.CartRepository.GetCartByUserId(command.UserId));
-        _unitOfWork.OrderRepository.DeleteRange(await _unitOfWork.OrderRepository.OrdersByUserAsync(command.UserId));
         _unitOfWork.UserRepository.Delete(user);
 
         await _unitOfWork.SaveChangesAsync(cancellationToken);
